+++ src/main/java/com/edu/certus/alumnos/repository/AlumnosRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.edu.certus.alumnos.entity.AlumnosEntity;
 
 @Repository
-public interface AlumnosRepository extends JpaRepository<AlumnosEntity, Long>{
+public interface AlumnosRepository extends JpaRepository<AlumnosEntity, Integer>{
 
 }
